% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_slope_part}
\alias{all_slope_part}
\title{MCMC tibble with regression parameters for HDAC6, KDM6A, MAOA, mPAS, NONO, P53, PTEN, STAG2}
\format{
A data frame with 6 variables: \code{x_intercept}, \code{mean_slope},
\code{sd_slope}, \code{partials_mean}, \code{partials_sd},
\code{clonal_mark}
}
\usage{
all_slope_part
}
\description{
MCMC tibble with regression parameters for HDAC6, KDM6A, MAOA, mPAS, NONO, P53, PTEN, STAG2
}
\keyword{datasets}
