% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fission_funcs.R
\name{calc_prob_below}
\alias{calc_prob_below}
\title{Calculate probability of chosen patch size or smaller}
\usage{
calc_prob_below(rho, age, size_measured)
}
\arguments{
\item{rho}{fission rate}

\item{age}{patient Age}

\item{size_measured}{patch size of interest}
}
\value{
Probability of patch size or less
}
\description{
Calculate probability of chosen patch size or smaller
}
