% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fission_funcs.R
\name{clonal_mark_fission_outpat}
\alias{clonal_mark_fission_outpat}
\title{Fit model to patch sizes}
\usage{
clonal_mark_fission_outpat(pat_data, num_chains = 4, multicore = T)
}
\arguments{
\item{pat_data}{data frame containing all clone info.}

\item{num_chains}{Number of MCMC chains to run default 4.}

\item{multicore}{Use multicore boolean, defaults to TRUE.}
}
\value{
A list with two MCMC tibbles one with per patient info and one with population level info
}
\description{
Fit model to patch sizes
}
