% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_part_funcs.R
\name{clonal_mark_regression}
\alias{clonal_mark_regression}
\title{Fit model to monoclonals and partials}
\usage{
clonal_mark_regression(
  count_mono,
  count_part,
  crypts_counted,
  Age,
  num_chains = 4,
  multicore = T
)
}
\arguments{
\item{count_mono}{Vector of monoclonal counts.}

\item{count_part}{Vector of partial clone counts.}

\item{crypts_counted}{Vector of crypt counts.}

\item{Age}{Vector of Ages.}

\item{num_chains}{Number of MCMC chains to run default 4.}

\item{multicore}{Use multicore boolean, defaults to TRUE.}
}
\value{
A tibble with the MCMC output of monoclonal and partial clone model fit
}
\description{
Fit model to monoclonals and partials
}
