% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_slides_pat.R
\name{convertSlidesToPatient}
\alias{convertSlidesToPatient}
\title{Process tibble with slide level info and sum to get patient level info}
\usage{
convertSlidesToPatient(full_part_all)
}
\arguments{
\item{full_part_all}{tibble with slide data.}
}
\value{
tibble with patient level data
}
\description{
Process tibble with slide level info and sum to get patient level info
}
