% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fission_marks}
\alias{fission_marks}
\title{MCMC tibble with fission population estimates for NONO, P53, mPAS, MAOA, Stag2}
\format{
A data frame with 3 variables: \code{pop_mu}, \code{pop_sd},
\code{clonal_mark}
\describe{
\item{pop_mu}{Mean fission of population}
\item{pop_sd}{Standard deviation of fission in population}
\item{clonal_mark}{One of several clonal marks}
}
}
\usage{
fission_marks
}
\description{
MCMC tibble with fission population estimates for NONO, P53, mPAS, MAOA, Stag2
}
\keyword{datasets}
