% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_part_funcs.R
\name{plot_clone_data}
\alias{plot_clone_data}
\title{Plot monoclonals and partials and fit}
\usage{
plot_clone_data(
  pat_data,
  col_mono = "steelblue4",
  col_part = "steelblue2",
  mcmc_fit = NULL,
  with_partials = F
)
}
\arguments{
\item{pat_data}{tibble with patient data.}

\item{col_mono}{Colour for monoclonal}

\item{col_part}{Colour for partials}

\item{mcmc_fit}{Include fit}

\item{with_partials}{Include partials}
}
\value{
grid plot
}
\description{
Plot monoclonals and partials and fit
}
