% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_part_funcs.R
\name{plot_clone_only_part}
\alias{plot_clone_only_part}
\title{Plot partials}
\usage{
plot_clone_only_part(pat_data, col_part = "steelblue2")
}
\arguments{
\item{pat_data}{tibble with patient data.}

\item{col_part}{Colour for partials}
}
\value{
plot
}
\description{
Plot partials
}
