% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_part_funcs.R
\name{plot_full_part_sep}
\alias{plot_full_part_sep}
\title{Plot monoclonals and partials separately}
\usage{
plot_full_part_sep(pat_data, col_mono = "steelblue4", col_part = "steelblue2")
}
\arguments{
\item{pat_data}{tibble with patient data.}

\item{col_mono}{Colour for monoclonal}

\item{col_part}{Colour for partials}
}
\value{
grid plot
}
\description{
Plot monoclonals and partials separately
}
