% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regr_maoa_mpas}
\alias{regr_maoa_mpas}
\title{MCMC tibble with regression parameters for mPAS and MAOA}
\format{
A data frame with 6 variables: \code{x_intercept}, \code{mean_slope},
\code{sd_slope}, \code{partials_mean}, \code{partials_sd},
\code{clonal_mark}
}
\usage{
regr_maoa_mpas
}
\description{
MCMC tibble with regression parameters for mPAS and MAOA
}
\keyword{datasets}
