% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHClones-package.R
\docType{package}
\name{RHClones-package}
\alias{RHClones-package}
\alias{RHClones}
\title{The 'RHClones' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
